NXP backend: Do not infer format for unknown nodes.
